Adding longer legislation proposals titles
Many of the titles that come in the file we have to import are longer than 80
characters, which is the maximum we allow.

The validation of the length of the title is based on the title_max_length
class method that is defined in the concern measurable. This method retrieves
the "limit" that may have been defined in the database and if it does not exist,
it defines by default a maximum of 80 characters.

The first step is to remove the database restriction.

The second step could be to overwrite the title_max_length method to modify
the default value, but with this solution we would be modifying the default
value not only of the title of the legislative proposals but also of all the
models that include the Measurable concern. So we have opted to overwrite
the model validation and instead of using title_max_length, we use the hardcoded
value of 140, which is the maximum value that exists for the titles of a
proposal in the file we have to import.

As we now allow longer titles in legislation processes, we also need to update
the css so that the title in the header is displayed correctly in the
legislation proposals index page.