Only render summary if present for legislation process
Currently the summary field and the description field are not mandatory.

The file we have to import has only a "welcome_message" field which is
too long to be included in the "summary" so we add it to the "description".

As this text has no line breaks, the "first_paragraph" method does not
work as expected and the whole description appears and is not visually
acceptable.

Rather than start adding line breaks to the file we have to import, it seems a
good solution to only show the summary field when it is actually reported.