Fix menu on small screens
The aspect ratio of the header logo is 1.6:1 instead of the 3.25:1 which
is the default one in Consul.

This made the menu look broken on small screens when users aren't
identified. While this has been fixed in Consul's master branch commit
045ac648d, we are increasing the size of the logo as a workaround until
Consul 2.0.0 is released and we upgrade to that version.