Allow linking a Document with a Legislation::Proposal
When we create a document as an administrator, we are allowing to relate it
to a Legislation::Proposal through its identifier.

Right now we only need to add the proposal ID in the form, since in the
controller we are already assuming that the documentable_type will always
be Legislation::Proposal.

If at some point we want to use more modules, it may be a good time to add
a select in the form to define the documentable_type. For now, as it is the
only module used, I think we should keep it as it is.