Render information text for comments when its commentable is hidden
Currently in consul links generated with polymorphic_path for a comment that
has its commentable hidden are not generated correctly.

We have seen that this patch is being applied to the comments index in
the administration section.

We apply this same patch in the moderation section to avoid adding
an invalid link.

A issue is created in Consul to analyse and resolve this problem.