Avoid duplicate <h1> tags on homepage
On production, the Junta de Castilla y León is using the site name as
the title of the homepage header. This results in two consecutive <h1>
tags with the same text.

So we're using a <div> tag on the homepage instead.

Note we're copying the original styles, which means will have to keep
them up to date when the original styles change.